Updated project location on GitHub
The former project name was too long, which caused issues with repeated arguments in data validation (invalid argument error). For the same reason, it is preferable to clone the project into a folder which is not too deep in the file system and without spaces, e.g. "D:\workspace\JSR-272-Eval-Kit" rather than "C:\Documents and Settings\All Users\Bureau\JSR-272-Eval-Kit"